# openapi.yml
openapi: 3.0.3
info:
  title: Fauna JavaScript sample app
  version: 1.0.0
  description: |
    HTTP API reference documentation for the [Fauna JavaScript sample
    app](https://github.com/fauna/js-sample-app).

    The app uses Fauna and Node.js to create HTTP API endpoints for a sample
    e-commerce store. You can use the endpoints to  manage customers,
    products, and orders for the store.

    For an overview and more details, see the
    [README](https://github.com/fauna/js-sample-app).

    ## Authentication

    Requests to the endpoints require no authentication. Internally, requests to
    Fauna are authenticated using a Fauna key's secret. You pass the secret
    using the `FAUNA_SECRET` environment variable in `.env`.

    For setup instructions, see the
    [README](https://github.com/fauna/js-sample-app).
servers:
  - url: "http://localhost:8000"
    description: Local development server
...
components:
  schemas:
...
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          description: "Unique, auto-generated ID for the customer."
        name:
          type: string
          description: Customer name.
        email:
          type: string
          format: email
          description: Customer's email address. Must be unique.
        address:
          type: object
          description: Customer's mailing address.
          properties:
            street:
              type: string
              description: Street of the customer's mailing address.
            city:
              type: string
              description: City of the customer's mailing address.
            state:
              type: string
              description: State of the customer's mailing address.
            postalCode:
              type: string
              description: Postal code of the customer's mailing address.
            country:
              type: string
              description: Country of the customer's mailing address.
      example:
        id: "111"
        name: John Doe
        email: john.doe@example.com
        address:
          street: 123 Main St
          city: San Francisco
          state: CA
          postalCode: "12345"
          country: United States
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          description: "Unique, auto-generated ID for the product."
        name:
          type: string
          description: Name of the product.
        price:
          type: number
          description: Price of the product in USD cents.
        description:
          type: string
          description: Description of the product.
        stock:
          type: integer
          description: Quantity of the product available in stock.
        category:
          type: object
          description: Category to which the product belongs.
          properties:
            id:
              type: string
              description: "Unique, auto-generated ID for the category."
            name:
              type: string
              description: Name of the category.
            description:
              type: string
              description: Description of the category.
      example:
        id: "123"
        name: The Old Man and the Sea
        price: 899
        description: A book by Ernest Hemingway
        stock: 10
        category:
          id: "123"
          name: books
          description: Bargain books!
...
    OrderRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - cart
            - processing
            - shipped
            - delivered
          description: |
            Status of the order.

            An order's status can only transition as follows:

            `cart` → `processing` → `shipped` → `delivered`

            Internally, the endpoint uses the `validateOrderStatusTransition()`
            user-defined function (UDF) to ensure an order's  `status` change is
            allowed.
        payment:
          type: object
          default: {}
          description: |
            Arbitrary object containing payment information.

            You can only update payment information if the `status` is `cart`.
            Adding payment information automatically transitions the order's
            `status` to `processing`.
      example:
        status: processing
    OrderPageResponse:
      type: object
      properties:
        results:
          type: array
          description: Page of orders.
          items:
            $ref: "#/components/schemas/OrderResponse"
        nextToken:
          type: string
          description: |
            Pagination token. Pass the token to the endpoint's `nextToken` query
            parameter to get the next page of results.
      example:
        results:
          - id: "123"
            payment: {}
            createdAt: "2099-08-09T16:43:07.153186Z"
            status: cart
            total: 899
            items:
              - product:
                  id: "123"
                  name: For Whom the Bell Tolls
                  price: 899
                  description: A book by Ernest Hemingway
                  stock: 10
                  category:
                    id: "123"
                    name: books
                    description: Bargain books!
                quantity: 1
            customer:
              id: "111"
              name: John Doe
              email: john.doe@example.com
              address:
                street: 123 Main St
                city: San Francisco
                state: CA
                postalCode: "12345"
                country: United States
        nextToken: hdW...
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          description: "Unique, auto-generated ID for the order."
        payment:
          type: object
          description: Payment details for the order (currently empty).
        createdAt:
          type: string
          format: date-time
          type: object
          description: ISO-8601-formatted timestamp of when the order was created.
        status:
          type: string
          description: Current status of the order.
          enum:
            - cart
            - processing
            - shipped
            - delivered
        total:
          type: integer
          description: Total cost of the order.
        items:
          type: array
          description: Items in the order.
          items:
            $ref: "#/components/schemas/OrderItem"
        customer:
          $ref: "#/components/schemas/CustomerResponse"
      example:
        id: "123"
        payment: {}
        createdAt: "2099-08-09T16:43:07.153186Z"
        status: cart
        total: 899
        items:
          - product:
              id: "123"
              name: For Whom the Bell Tolls
              price: 899
              description: A book by Ernest Hemingway
              stock: 10
              category:
                id: "123"
                name: books
                description: Bargain books!
            quantity: 1
        customer:
          id: "111"
          name: John Doe
          email: john.doe@example.com
          address:
            street: 123 Main St
            city: San Francisco
            state: CA
            postalCode: "12345"
            country: United States
    OrderItem:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/ProductResponse"
        quantity:
          type: integer
          description: Quantity of the product in the order.
      example:
        product:
          id: "123"
          name: For Whom the Bell Tolls
          price: 899
          description: A book by Ernest Hemingway
          stock: 7
          category:
            id: "123"
            name: books
            description: Bargain books!
          quantity: 1
    OrderItemRequest:
      type: object
      required:
        - productName
        - quantity
      properties:
        productName:
          type: string
          description: |
            Name of an existing product to add to the order.
        quantity:
          type: integer
          description: >
            Quantity of the product to add to the order.

            Must be a non-negative integer. The quantity must be less than the
            product's `stock`.

            Internally, the `OrderItem` collection uses a check constraint to
            ensure `quantity` is a positive integer.
      example:
        productName: The Old Man and the Sea
        quantity: 1
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message.
      example:
        message: Invalid request parameter
paths:
...
  "/customers/{id}/cart":
    get:
      summary: Get a customer's cart.
      tags:
        - Orders
      description: |
        Get the details of a customer's cart order.
      parameters:
        - description: ID of the customer.
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: No customer with id
                  value:
                    message: No customer with id '111' exists.
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v http://localhost:8000/customers/111/cart
    post:
      summary: Get or create a customer's cart.
      tags:
        - Orders
      description: |
        Get the details of a customer's cart. If no cart exists, create one with
        no items.

        A customer's cart is an order with the `cart` status. A customer can
        have one cart at a time.

        To add items to a customer's cart, use the `POST
        /customers/{id}/cart/item` endpoint.

        Internally, the endpoint uses the `getOrCreateCart()` user-defined
        function (UDF) to conditionally get or create the customer's cart.
      parameters:
        - description: ID of the customer.
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: No customer with id
                  value:
                    message: No customer with id '111' exists.
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v -X POST http://localhost:8000/customers/111/cart
  "/customers/{id}/cart/item":
    post:
      summary: Add an item to a customer's cart.
      tags:
        - Orders
      description: |
        Add an item to a customer's cart. If the order already contains the
        item's product, update the item's quantity.

        If the customer has no cart, the request create ones.

        The request returns an item. An item links a product, with a quantity,
        to an order. An order contains zero or more items.

        Internally, the endpoint uses the `createOrUpdateCartItem()`
        user-defined function (UDF) to conditionally update the item's quantity
        and create a new cart. The UDF also handles various validations and
        error cases for the request.
      parameters:
        - description: ID of the customer.
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-product:
                  summary: Invalid product.
                  value:
                    message: Product must be a non empty string.
                invalid-quantity:
                  summary: Invalid quantity.
                  value:
                    message: Quantity must be a positive integer.
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: No customer with id
                  value:
                    message: No customer with id '111' exists.
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v -X POST http://localhost:8000/customers/111/cart/item \
              --header 'content-type: application/json' \
              --data '{"productName":"The Old Man and the Sea","quantity":1}'
  "/orders/{id}":
    get:
      summary: Get an order.
      tags:
        - Orders
      description: Get the details of an order by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the order to fetch.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v http://localhost:8000/orders/123
    patch:
      summary: Update an order.
      tags:
        - Orders
      description: >
        Update the details of an order by its ID.


        Changing the order's `status` to `processing` runs a checkout on the
        order.


        Internally, the endpoint uses the `checkout()` user-defined function

        (UDF) to validate orders during checkout. The UDF:

         * Verifies the order is in a `processing` status.
         * Ensures the order only adds valid, existing products as items.
         * Verifies that the order does not reduce any product's stock below zero.
         * Ensures the order has a non-null `payment`.
      parameters:
        - name: id
          in: path
          description: ID of the order to update.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-payment-change-a:
                  summary: Payment change not allowed.
                  value:
                    message: >-
                      Payment method may only be updated before the order has
                      been placed.
                invalid-payment-change-b:
                  summary: Payment change not allowed.
                  value:
                    message: >-
                      Can not update payment information after an order has been
                      placed.
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: Unauthorized
                  value:
                    message: >-
                      Unauthorized. Set the `FAUNA_SECRET` env var to a valid
                      Fauna key's secret.
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: No order with id
                  value:
                    message: No order with id '111' exists.
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example-1:
                  summary: Internal server error
                  value:
                    message: An unexpected error occurred
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v -X PATCH http://localhost:8000/orders/123 \
              --header 'content-type: application/json' \
              --data '{"status":"processing"}'
  "/customers/{id}/orders":
    get:
      summary: Get a customer's orders.
      tags:
        - Orders
      description: Get the details of a customer's orders using the customer's ID.
      parameters:
        - description: ID of the customer.
          in: path
          name: id
          required: true
          schema:
            type: string
        - description: |
            Accepts a pagination token. The endpoint's responses
            return these tokens in the `nextToken` property.
          in: query
          name: nextToken
          schema:
            type: string
        - description: Number of products to return per page.
          in: query
          name: pageSize
          schema:
            default: 10
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-pagesize:
                  summary: Invalid pageSize.
                  value:
                    message: Page size must be a positive integer or be omitted.
                invalid-next-token:
                  summary: Invalid nextToken.
                  value:
                    message: Next token must be a string or be omitted.
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                example:
                  summary: No customer with id
                  value:
                    message: No customer with id '111' exists.
      x-codeSamples:
        - lang: curl
          source: |-
            curl -v 'http://localhost:8000/customers/111/orders?pageSize=1'